health: #152 moved lexer code from @entity-space/core to @entity-space/lexer
