+++ packages/core/src/lib/criteria/parser/token-parser/token-parser.type.ts
@@ -1,4 +1,5 @@
+import { Token } from "@entity-space/lexer";
 import { Criterion } from "../../criterion";
-import { Token } from "../token.contract";
 
+// [todo] rename to CriteriaTokenParser
 export type TokenParser = Generator<undefined | (() => Criterion), false | (() => Criterion), Token>;
