+++ packages/lexer/src/index.ts
@@ -1,1 +1,4 @@
-export * from "./lib/lexer";
+export * from "./lib/lex.fn";
+export * from "./lib/token-type.enum";
+export * from "./lib/token.contract";
+export * from "./lib/token.fn";
