+++ packages/lexer/src/lib/lex.fn.ts
