--- packages/lexer/src/lib/lexer.ts
@@ -1,3 +1,0 @@
-export function lexer(): string {
-    return "lexer";
-}
