+++ packages/lexer/src/lib/scan-number.fn.ts
