+++ packages/lexer/src/lib/scan-string.fn.ts
