+++ packages/lexer/src/lib/scan-symbol.fn.ts
