+++ packages/lexer/src/lib/token-type.enum.ts
