+++ packages/lexer/src/lib/token.fn.ts
